hal/armv7m & armv7a: common pmap interface
more:
 - remove errno from interrupts

JIRA: PD-210
