+++ hal/armv7a/arch/pmap.h
@@ -13,8 +13,10 @@
  * %LICENSE%
  */
 
-#ifndef _HAL_PMAP_H_
-#define _HAL_PMAP_H_
+#ifndef _HAL_PMAP_ARMV7A_H_
+#define _HAL_PMAP_ARMV7A_H_
+
+#include "types.h"
 
 /* Predefined virtual adresses */
 #define VADDR_KERNEL   0xc0000000   /* base virtual address of kernel space */
@@ -51,10 +53,7 @@
 #define PAGE_KERNEL_STACK    (5 << 4)
 #define PAGE_KERNEL_HEAP     (6 << 4)
 
-
 #ifndef __ASSEMBLY__
-
-#include <arch/types.h>
 
 /* Structure describing page - its should be aligned to 2^N boundary */
 typedef struct _page_t {
@@ -75,49 +74,6 @@ typedef struct _pmap_t {
 	void *pmapv;
 	page_t *pmapp;
 } pmap_t;
-
-
-static inline int pmap_belongs(pmap_t *pmap, void *addr)
-{
-	return addr >= pmap->start && addr < pmap->end;
-}
-
-
-extern int pmap_create(pmap_t *pmap, pmap_t *kpmap, page_t *p, void *vaddr);
-
-
-extern addr_t pmap_destroy(pmap_t *pmap, int *i);
-
-
-extern void pmap_switch(pmap_t *pmap);
-
-
-extern int pmap_enter(pmap_t *pmap, addr_t addr, void *vaddr, int attrs, page_t *alloc);
-
-
-/* Function removes mapping for given address */
-extern int pmap_remove(pmap_t *pmap, void *vaddr);
-
-
-extern addr_t pmap_resolve(pmap_t *pmap, void *vaddr);
-
-
-extern int pmap_getPage(page_t *page, addr_t *addr);
-
-
-/* Function return character marker for page flags */
-extern char pmap_marker(page_t *p);
-
-
-/* Function allocates page tables for kernel space */
-extern int _pmap_kernelSpaceExpand(pmap_t *pmap, void **start, void *end, page_t *dp);
-
-
-extern int pmap_segment(unsigned int i, void **vaddr, size_t *size, int *prot, void **top);
-
-
-extern void _pmap_init(pmap_t *pmap, void **start, void **end);
-
 
 #endif
