+++ hal/armv7m/imxrt/interrupts.c
@@ -15,12 +15,9 @@
 
 #include "../../interrupts.h"
 #include "../../spinlock.h"
-#include "../../cpu.h"
 #include "config.h"
-#include "pmap.h"
 
 #include "../../proc/userintr.h"
-#include "../../include/errno.h"
 
 
 #define _intr_add(list, t) \
@@ -100,7 +97,7 @@ int hal_interruptsSetHandler(intr_handler_t *h)
 	spinlock_ctx_t sc;
 
 	if (h == NULL || h->f == NULL || h->n >= SIZE_INTERRUPTS)
-		return -EINVAL;
+		return -1;
 
 	hal_spinlockSet(&interrupts.spinlock, &sc);
 	h->got = hal_cpuGetGot();
@@ -113,7 +110,7 @@ int hal_interruptsSetHandler(intr_handler_t *h)
 	}
 	hal_spinlockClear(&interrupts.spinlock, &sc);
 
-	return EOK;
+	return 0;
 }
 
 
@@ -122,7 +119,7 @@ int hal_interruptsDeleteHandler(intr_handler_t *h)
 	spinlock_ctx_t sc;
 
 	if (h == NULL || h->f == NULL || h->n >= SIZE_INTERRUPTS)
-		return -EINVAL;
+		return -1;
 
 	hal_spinlockSet(&interrupts.spinlock, &sc);
 	_intr_remove(&interrupts.handlers[h->n], h);
@@ -132,7 +129,7 @@ int hal_interruptsDeleteHandler(intr_handler_t *h)
 
 	hal_spinlockClear(&interrupts.spinlock, &sc);
 
-	return EOK;
+	return 0;
 }
