+++ hal/armv7m/pmap.c
@@ -13,7 +13,8 @@
  * %LICENSE%
  */
 
-#include "pmap.h"
+#include "../pmap.h"
+#include <arch/cpu.h>
 
 /* Linker symbols */
 extern unsigned int _end;
@@ -22,9 +23,28 @@ extern unsigned int __bss_start;
 extern void *_init_vectors;
 
 
+/* Function creates empty page table */
+int pmap_create(pmap_t *pmap, pmap_t *kpmap, page_t *p, void *vaddr)
+{
+	return 0;
+}
+
+
+addr_t pmap_destroy(pmap_t *pmap, int *i)
+{
+	return 0;
+}
+
+
 void pmap_switch(pmap_t *pmap)
 {
 	return;
+}
+
+
+int pmap_enter(pmap_t *pmap, addr_t pa, void *vaddr, int attr, page_t *alloc)
+{
+	return 0;
 }
 
 
@@ -34,7 +54,25 @@ int pmap_remove(pmap_t *pmap, void *vaddr)
 }
 
 
-int pmap_enter(pmap_t *pmap, addr_t pa, void *vaddr, int attr, page_t *alloc)
+addr_t pmap_resolve(pmap_t *pmap, void *vaddr)
+{
+	return (addr_t)vaddr;
+}
+
+
+int pmap_getPage(page_t *page, addr_t *addr)
+{
+	return 0;
+}
+
+
+char pmap_marker(page_t *p)
+{
+	return 0;
+}
+
+
+int _pmap_kernelSpaceExpand(pmap_t *pmap, void **start, void *end, page_t *dp)
 {
 	return 0;
 }
@@ -49,13 +87,6 @@ int pmap_segment(unsigned int i, void **vaddr, size_t *size, int *prot, void **t
 	*vaddr = (void *)&_end;
 	*size = (((size_t)(*top) + SIZE_PAGE - 1) & ~(SIZE_PAGE - 1)) - (size_t)&_end;
 
-	return 0;
-}
-
-
-/* Function creates empty page table */
-int pmap_create(pmap_t *pmap, pmap_t *kpmap, page_t *p, void *vaddr)
-{
 	return 0;
 }
